+++ CFI/buffer.d
@@ -10,7 +10,7 @@ File create_block(string filename) {
   return create_file(filename);
 }
 
-void write_block(string table_name, ulong block_id, string content) {
+void write_block(string table_name, uint block_id, string content) {
   writeln(table_name, block_id, content);
 
   string file_name = format("%s.%s", table_name, RECORD_EXTENSION);
